Users, Roles and Rights are now functioning though security has not been fully integrated.  Currently it is only being used to a limited degree to determine which menus to display and, perhaps in the case of Users, Roles and Rights, what can be done within a controller or two.
Need to check on the proper callback for an on_create (can't remember and I'm offline at the moment) so that attempts can be set to zero automatically on the User object rather than having to kludge it into the controller.  The model is better because it protects from rails c tinkering.

Menus are now based on the user rights.  Rights are, of course, governed exclusively by roles.

The Admin right trumps all.  Currently this is the only right and role pre-populated in the system.  There will definitely be others since I'm adding code that looks for "commentators", "analysts" and "viewers."

Active model selection in the navbar has been fixed.  Didn't realize that it was broken.

Created a couple of Safari specific PNG placeholders so that it Safari stops spamming the logs.

If in Dev mode, SSL is not enforced and secure cookies are also not enforced.